requirements.txt
Flask==3.0.2
pandas==2.2.1
openpyxl==3.1.2
jinja2==3.1.3
weasyprint==52.5
python-dotenv==1.0.1
openai==1.55.3 